+++ lib/presentation/search_recipes/search_recipes_screen.dart
@@ -4,8 +4,6 @@ import 'package:recipe_app/core/error/ui_state.dart';
 // Project imports:
 import 'package:recipe_app/core/ui/color_style.dart';
 import 'package:recipe_app/core/ui/text_style.dart';
-import 'package:recipe_app/data/mocks/mock_recipe_data_source_impl.dart';
-import 'package:recipe_app/data/repository/recipe_repository_impl.dart';
 import 'package:recipe_app/presentation/component/filter_search_bottom_sheet/filter_search_bottom_sheet.dart';
 import 'package:recipe_app/presentation/component/filter_search_bottom_sheet/filter_search_state.dart';
 import 'package:recipe_app/presentation/component/image_component/app_image.dart';
@@ -19,17 +17,6 @@ class SearchRecipesScreen extends StatefulWidget {
   // List<Recipe> _allRecipes = [];
   const SearchRecipesScreen({super.key, required this.viewModel});
 
-  factory SearchRecipesScreen.withMock() {
-    return SearchRecipesScreen(
-      viewModel: SearchRecipesViewModel(
-        RecipeRepositoryImpl(
-          remoteDataSource: null,
-          localDataSource: MockRecipeDataSourceImpl(),
-        ),
-      ),
-    );
-  }
-
   @override
   State<SearchRecipesScreen> createState() => _SearchRecipesScreenState();
 }
@@ -39,6 +26,12 @@ class _SearchRecipesScreenState extends State<SearchRecipesScreen> {
   void initState() {
     super.initState();
     widget.viewModel.load();
+  }
+
+  @override
+  void dispose() {
+    widget.viewModel.dispose();
+    super.dispose();
   }
 
   @override
